Speed up running the tests of the IO module from Maven
The `mvn test` command, when run in the IO module, now takes 15 seconds instead of 5 minutes.
This has been achieved by moving many of the tests to the verification phase instead of the test phase, by running them with the failsafe plugin instead of surefire.
Also, the tests are now executed in parallel.
