+++ community/io/pom.xml
@@ -120,6 +120,13 @@ the relevant Commercial Agreement.
           </jvmArgs>
         </configuration>
       </plugin>
+      <plugin>
+        <artifactId>maven-surefire-plugin</artifactId>
+        <configuration>
+          <parallel>methods</parallel>
+          <perCoreThreadCount>2</perCoreThreadCount>
+        </configuration>
+      </plugin>
     </plugins>
   </build>
